requirements.txt
fastapi
elasticsearch
langchain
pydantic
numpy
docling
pypdf
requests
pillow
aiohttp
dependency-injector
google-genai